+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Support temorary songs that can exist only in playlist.
 - Add option to directly play file.
 - Add option to sort the playlist.
+- Add aliases for control messages.
 
 ## v0.4.0
 - Add all options to settings including help